Fix dummy FSM used in tests
`0` has a special meaning as both state and event, so actual states and events
must start from `1`.

The "NULL event" - means there are no events pending.
The "NULL state" is used in the transition table. Target state '0' means there
is no transition associated with given state and event.

This is solved by generating FSM code from the model, so the risk of making a
mistake exists only when coding FSM states, events and tables manually.